@@ -1,17 +1,17 @@
 # frozen_string_literal: true
 
-require_relative "lib/phony_gem/version"
+require_relative 'lib/phony_gem/version'
 
 Gem::Specification.new do |spec|
-  spec.name = "phony_gem"
+  spec.name = 'phony_gem'
   spec.version = PhonyGem::VERSION
-  spec.authors = ["Felix"]
-  spec.email = ["zen9.felix@gmail.com"]
+  spec.authors = ['Felix']
+  spec.email = ['zen9.felix@gmail.com']
 
-  spec.summary = "Write a short summary, because RubyGems requires one."
-  spec.description = "Write a longer description or delete this line."
-  spec.homepage = "https://xxx.com"
-  spec.required_ruby_version = ">= 2.6.0"
+  spec.summary = 'Write a short summary, because RubyGems requires one.'
+  spec.description = 'Write a longer description or delete this line.'
+  spec.homepage = 'https://xxx.com'
+  spec.required_ruby_version = '>= 2.7.0'
 
   # Specify which files should be added to the gem when it is released.
   # The `git ls-files -z` loads the files in the RubyGem that have been added into git.
@@ -20,13 +20,14 @@ Gem::Specification.new do |spec|
       (f == __FILE__) || f.match(%r{\A(?:(?:bin|test|spec|features)/|\.(?:git|travis|circleci)|appveyor)})
     end
   end
-  spec.bindir = "exe"
+  spec.bindir = 'exe'
   spec.executables = spec.files.grep(%r{\Aexe/}) { |f| File.basename(f) }
-  spec.require_paths = ["lib"]
+  spec.require_paths = ['lib']
 
   # Uncomment to register a new dependency of your gem
   # spec.add_dependency "example-gem", "~> 1.0"
 
   # For more information and examples about making a new gem, check out our
   # guide at: https://bundler.io/guides/creating_gem.html
+  spec.metadata['rubygems_mfa_required'] = 'true'
 end
