@@ -1,24 +1,27 @@
-require "optparse"
-require "tmpdir"
-require "tempfile"
-require "pathname"
+# frozen_string_literal: true
 
-require_relative "../bundix"
-require_relative "shell_nix_context"
+require 'English'
+require 'optparse'
+require 'tmpdir'
+require 'tempfile'
+require 'pathname'
+
+require_relative '../bundix'
+require_relative 'shell_nix_context'
 
 class Bundix
   class CommandLine
     DEFAULT_OPTIONS = {
-      ruby: "ruby",
-      bundle_pack_path: "vendor/bundle",
-      gemfile: "Gemfile",
-      lockfile: "Gemfile.lock",
-      gemset: "gemset.nix",
-      project: File.basename(Dir.pwd),
-    }
+      ruby: 'ruby',
+      bundle_pack_path: 'vendor/bundle',
+      gemfile: 'Gemfile',
+      lockfile: 'Gemfile.lock',
+      gemset: 'gemset.nix',
+      project: File.basename(Dir.pwd)
+    }.freeze
 
     def self.run
-      self.new.run
+      new.run
     end
 
     def initialize
@@ -37,54 +40,57 @@ class Bundix
 
     def parse_options
       op = OptionParser.new do |o|
-        o.on "-m", "--magic", "lock, pack, and write dependencies" do
+        o.on '-m', '--magic', 'lock, pack, and write dependencies' do
           options[:magic] = true
         end
 
-        o.on "--ruby=#{options[:ruby]}", "ruby version to use for magic and init, defaults to latest" do |value|
+        o.on "--ruby=#{options[:ruby]}",
+             'ruby version to use for magic and init, defaults to latest' do |value|
           options[:ruby] = value
         end
 
-        o.on "--bundle-pack-path=#{options[:bundle_pack_path]}", "path to pack the magic" do |value|
+        o.on "--bundle-pack-path=#{options[:bundle_pack_path]}",
+             'path to pack the magic' do |value|
           options[:bundle_pack_path] = value
         end
 
-        o.on "-i", "--init", "initialize a new shell.nix for nix-shell (won't overwrite old ones)" do
+        o.on '-i', '--init',
+             "initialize a new shell.nix for nix-shell (won't overwrite old ones)" do
           options[:init] = true
         end
 
-        o.on "--gemset=#{options[:gemset]}", "path to the gemset.nix" do |value|
+        o.on "--gemset=#{options[:gemset]}", 'path to the gemset.nix' do |value|
           options[:gemset] = File.expand_path(value)
         end
 
-        o.on "--lockfile=#{options[:lockfile]}", "path to the Gemfile.lock" do |value|
+        o.on "--lockfile=#{options[:lockfile]}", 'path to the Gemfile.lock' do |value|
           options[:lockfile] = File.expand_path(value)
         end
 
-        o.on "--gemfile=#{options[:gemfile]}", "path to the Gemfile" do |value|
+        o.on "--gemfile=#{options[:gemfile]}", 'path to the Gemfile' do |value|
           options[:gemfile] = File.expand_path(value)
         end
 
-        o.on "-d", "--dependencies", "include gem dependencies (deprecated)" do
-          warn "--dependencies/-d is deprecated because"
-          warn "dependencies will always be fetched"
+        o.on '-d', '--dependencies', 'include gem dependencies (deprecated)' do
+          warn '--dependencies/-d is deprecated because'
+          warn 'dependencies will always be fetched'
         end
 
-        o.on "-q", "--quiet", "only output errors" do
+        o.on '-q', '--quiet', 'only output errors' do
           options[:quiet] = true
         end
 
-        o.on "-l", "--lock", "generate Gemfile.lock first" do
+        o.on '-l', '--lock', 'generate Gemfile.lock first' do
           options[:lock] = true
         end
 
-        o.on "-v", "--version", "show the version of bundix" do
+        o.on '-v', '--version', 'show the version of bundix' do
           puts Bundix::VERSION
           exit
         end
 
-        o.on "--env", "show the environment in bundix" do
-          system("env")
+        o.on '--env', 'show the environment in bundix' do
+          system('env')
           exit
         end
       end
@@ -95,20 +101,19 @@ class Bundix
     end
 
     def handle_magic
-      ENV["BUNDLE_GEMFILE"] = options[:gemfile]
+      ENV['BUNDLE_GEMFILE'] = options[:gemfile]
 
-      if options[:magic]
-        fail unless system(
-          Bundix::NIX_SHELL, "-p", options[:ruby],
-          "bundler.override { ruby = #{options[:ruby]}; }",
-          "--command", "bundle lock --lockfile=#{options[:lockfile]}"
-        )
-        fail unless system(
-          Bundix::NIX_SHELL, "-p", options[:ruby],
-          "bundler.override { ruby = #{options[:ruby]}; }",
-          "--command", "bundle pack --all --path #{options[:bundle_pack_path]}"
-        )
-      end
+      return unless options[:magic]
+      raise unless system(
+        Bundix::NIX_SHELL, '-p', options[:ruby],
+        "bundler.override { ruby = #{options[:ruby]}; }",
+        '--command', "bundle lock --lockfile=#{options[:lockfile]}"
+      )
+      raise unless system(
+        Bundix::NIX_SHELL, '-p', options[:ruby],
+        "bundler.override { ruby = #{options[:ruby]}; }",
+        '--command', "bundle pack --all --path #{options[:bundle_pack_path]}"
+      )
     end
 
     def shell_nix_context
@@ -116,22 +121,22 @@ class Bundix
     end
 
     def handle_lock
-      if options[:lock]
-        lock = !File.file?(options[:lockfile])
-        lock ||= File.mtime(options[:gemfile]) > File.mtime(options[:lockfile])
-        if lock
-          ENV.delete("BUNDLE_PATH")
-          ENV.delete("BUNDLE_FROZEN")
-          ENV.delete("BUNDLE_BIN_PATH")
-          system("bundle", "lock")
-          fail "bundle lock failed" unless $?.success?
-        end
-      end
+      return unless options[:lock]
+
+      lock = !File.file?(options[:lockfile])
+      lock ||= File.mtime(options[:gemfile]) > File.mtime(options[:lockfile])
+      return unless lock
+
+      ENV.delete('BUNDLE_PATH')
+      ENV.delete('BUNDLE_FROZEN')
+      ENV.delete('BUNDLE_BIN_PATH')
+      system('bundle', 'lock')
+      raise 'bundle lock failed' unless $CHILD_STATUS.success?
     end
 
     def build_gemset
-      unless File.exist? "Gemfile.lock"
-        puts "missing Gemfile.lock"
+      unless File.exist? 'Gemfile.lock'
+        puts 'missing Gemfile.lock'
         return
       end
 
@@ -143,12 +148,12 @@ class Bundix
     end
 
     def save_gemset(gemset)
-      tempfile = Tempfile.new("gemset.nix", encoding: "UTF-8")
+      tempfile = Tempfile.new('gemset.nix', encoding: 'UTF-8')
       begin
         tempfile.write(object2nix(gemset))
         tempfile.flush
         FileUtils.cp(tempfile.path, options[:gemset])
-        FileUtils.chmod(0644, options[:gemset])
+        FileUtils.chmod(0o644, options[:gemset])
       ensure
         tempfile.close!
         tempfile.unlink
