@@ -1,5 +1,7 @@
+# frozen_string_literal: true
+
 class Bundix
-  class ShellNixContext < Struct.new(:project, :ruby, :gemfile, :lockfile, :gemset)
+  ShellNixContext = Struct.new(:project, :ruby, :gemfile, :lockfile, :gemset) do
     def self.from_hash(hash)
       new(*hash.values_at(*members))
     end
@@ -9,7 +11,7 @@ class Bundix
     end
 
     def path_for(file)
-      Nixer.serialize(Pathname(file).relative_path_from(Pathname("./")))
+      Nixer.serialize(Pathname(file).relative_path_from(Pathname('./')))
     end
 
     def gemfile_path
