@@ -1,7 +1,9 @@
+# frozen_string_literal: true
+
 class Bundix
   class Fetcher
-    def sh(*args, &block)
-      Bundix.sh(*args, &block)
+    def sh(...)
+      Bundix.sh(...)
     end
 
     def download(file, url)
@@ -9,9 +11,7 @@ class Bundix
       uri = URI(url)
       open_options = {}
 
-      unless uri.user
-        inject_credentials_from_bundler_settings(uri)
-      end
+      inject_credentials_from_bundler_settings(uri) unless uri.user
 
       if uri.user
         open_options[:http_basic_authentication] = [uri.user, uri.password]
@@ -20,10 +20,10 @@ class Bundix
       end
 
       begin
-        open(uri.to_s, "r", 0600, open_options) do |net|
-          File.open(file, "wb+") { |local|
+        URI.parse(uri.to_s).open('r', 0o600, open_options) do |net|
+          File.open(file, 'wb+') do |local|
             File.copy_stream(net, local)
-          }
+          end
         end
       rescue OpenURI::HTTPError => e
         # e.message: "403 Forbidden" or "401 Unauthorized"
@@ -33,62 +33,63 @@ class Bundix
     end
 
     def inject_credentials_from_bundler_settings(uri)
-      @bundler_settings ||= Bundler::Settings.new(Bundler.root + ".bundle")
+      # rubocop:disable all
+      @bundler_settings ||= Bundler::Settings.new(Bundler.root + '.bundle')
 
-      if val = @bundler_settings[uri.host]
-        uri.user, uri.password = val.split(":", 2)
-      end
+      return unless (val = @bundler_settings[uri.host])
+
+      uri.user, uri.password = val.split(':', 2)
     end
 
     def debrief_access_denied(host)
       print_error(
-        "Authentication is required for #{host}.\n" +
-        "Please supply credentials for this source. You can do this by running:\n" +
-        " bundle config packages.shopify.io username:password"
+        "Authentication is required for #{host}.\n" \
+        "Please supply credentials for this source. You can do this by running:\n" \
+        ' bundle config packages.shopify.io username:password'
       )
     end
 
     def print_error(msg)
       msg = "\x1b[31m#{msg}\x1b[0m" if $stdout.tty?
-      STDERR.puts(msg)
+      warn(msg)
     end
 
     def nix_prefetch_url(url)
-      dir = File.join(ENV["XDG_CACHE_HOME"] || "#{ENV["HOME"]}/.cache", "bundix")
+      dir = File.join(ENV['XDG_CACHE_HOME'] || "#{ENV['HOME']}/.cache", 'bundix')
       FileUtils.mkdir_p dir
-      file = File.join(dir, url.gsub(/[^\w-]+/, "_"))
+      file = File.join(dir, url.gsub(/[^\w-]+/, '_'))
 
       download(file, url) unless File.size?(file)
       return unless File.size?(file)
 
       sh(
         Bundix::NIX_PREFETCH_URL,
-        "--type", "sha256",
-        "--name", File.basename(url), # --name mygem-1.2.3.gem
-        "file://#{file}",             # file:///.../https_rubygems_org_gems_mygem-1_2_3_gem
-      ).force_encoding("UTF-8").strip
-    rescue => ex
-      puts ex
+        '--type', 'sha256',
+        '--name', File.basename(url), # --name mygem-1.2.3.gem
+        "file://#{file}" # file:///.../https_rubygems_org_gems_mygem-1_2_3_gem
+      ).force_encoding('UTF-8').strip
+    rescue StandardError => e
+      puts e
       nil
     end
 
     def nix_prefetch_git(uri, revision, submodules: false)
-      home = ENV["HOME"]
-      ENV["HOME"] = "/homeless-shelter"
+      home = ENV['HOME']
+      ENV['HOME'] = '/homeless-shelter'
 
       args = []
-      args << "--url" << uri
-      args << "--rev" << revision
-      args << "--hash" << "sha256"
-      args << "--fetch-submodules" if submodules
+      args << '--url' << uri
+      args << '--rev' << revision
+      args << '--hash' << 'sha256'
+      args << '--fetch-submodules' if submodules
 
       sh(NIX_PREFETCH_GIT, *args)
     ensure
-      ENV["HOME"] = home
+      ENV['HOME'] = home
     end
 
     def format_hash(hash)
-      sh(NIX_HASH, "--type", "sha256", "--to-base32", hash)[SHA256_32]
+      sh(NIX_HASH, '--type', 'sha256', '--to-base32', hash)[SHA256_32]
     end
 
     def fetch_local_hash(spec)
@@ -97,10 +98,10 @@ class Bundix
       filename = has_platform ? "#{name_version}-*" : name_version
 
       paths = spec.source.caches.map(&:to_s)
-      Dir.glob("{#{paths.join(",")}}/#{filename}.gem").each do |path|
+      Dir.glob("{#{paths.join(',')}}/#{filename}.gem").each do |path|
         if has_platform
           # Find first gem that matches the platform
-          platform = File.basename(path, ".gem")[(name_version.size + 1)..-1]
+          platform = File.basename(path, '.gem')[(name_version.size + 1)..]
           next unless spec.platform =~ platform
         end
 
@@ -133,8 +134,8 @@ class Bundix
       result = nix_prefetch_url(uri)
       return unless result
 
-      return result[SHA256_32], spec.platform&.to_s
-    rescue => e
+      [result[SHA256_32], spec.platform&.to_s]
+    rescue StandardError => e
       puts "ignoring error during fetching: #{e}"
       puts e.backtrace
       nil
@@ -152,14 +153,14 @@ class Bundix
       dep = Gem::Dependency.new(dependency.name, dependency.version)
       match_current_platform = false
       specs, _errors = Gem::SpecFetcher.new(sources).spec_for_dependency(dep, match_current_platform)
-      specs.each do |spec, source|
+      specs.each do |spec, _source|
         return spec if dependency.platform == spec.platform
       end
-      fail "Unable to find compatible rubygem source for #{dependency.platform.to_s}."
+      raise "Unable to find compatible rubygem source for #{dependency.platform}."
     end
   end
 
-  class Source < Struct.new(:spec, :fetcher)
+  Source = Struct.new(:spec, :fetcher) do
     def convert
       case spec.source
       when Bundler::Source::Rubygems
@@ -170,63 +171,64 @@ class Bundix
         convert_path
       else
         pp spec
-        fail "unknown bundler source"
+        raise 'unknown bundler source'
       end
     end
 
     def convert_path
       {
-        "version" => spec.version.to_s,
-        "source" => {
-          "type" => "path",
-          "path" => spec.source.path.to_s,
-        },
+        'version' => spec.version.to_s,
+        'source' => {
+          'type' => 'path',
+          'path' => spec.source.path.to_s
+        }
       }
     end
 
     def convert_rubygems
-      remotes = spec.source.remotes.map { |remote| remote.to_s.sub(/\/+$/, "") }
+      remotes = spec.source.remotes.map { |remote| remote.to_s.sub(%r{/+$}, '') }
       hash, platform = fetcher.fetch_local_hash(spec)
       remote, hash, platform = fetcher.fetch_remotes_hash(spec, remotes) unless hash
-      fail "couldn't fetch hash for #{spec.full_name}" unless hash
+      raise "couldn't fetch hash for #{spec.full_name}" unless hash
 
       version = spec.version.to_s
       nixspec = {
-        "version" => version,
-        "source" => {
-          "type" => "gem",
-          "remotes" => (remote ? [remote] : remotes),
-          "sha256" => hash,
-          "target" => platform,
-        },
+        'version' => version,
+        'source' => {
+          'type' => 'gem',
+          'remotes' => (remote ? [remote] : remotes),
+          'sha256' => hash,
+          'target' => platform
+        }
       }
       native_platform = Gem::Platform.new(platform)
       if native_platform.respond_to? :cpu
-        nixspec["source"]["targetCPU"] = native_platform.cpu
-        nixspec["source"]["targetOS"] = native_platform.os
+        nixspec['source']['targetCPU'] = native_platform.cpu
+        nixspec['source']['targetOS'] = native_platform.os
       end
       nixspec
     end
 
     def convert_git
-      revision = spec.source.options.fetch("revision")
-      uri = spec.source.options.fetch("uri")
-      submodules = !!spec.source.submodules
+      revision = spec.source.options.fetch('revision')
+      uri = spec.source.options.fetch('uri')
+      submodules = !spec.source.submodules.nil?
       output = fetcher.nix_prefetch_git(uri, revision, submodules: submodules)
       # FIXME: this is a hack, we should separate $stdout/$stderr in the sh call
-      hash = JSON.parse(output[/({[^}]+})\s*\z/m])["sha256"]
-      fail "couldn't fetch hash for #{spec.full_name}" unless hash
+      hash = JSON.parse(output[/({[^}]+})\s*\z/m])['sha256']
+      raise "couldn't fetch hash for #{spec.full_name}" unless hash
+
       puts "#{hash} => #{uri}" if $VERBOSE
 
       {
-        "version" => spec.version.to_s,
-        "source" => {
-          "type" => "git",
-          "url" => uri.to_s,
-          "rev" => revision,
-          "sha256" => hash,
-          "fetchSubmodules" => submodules,
-        },
+        'version' => spec.version.to_s,
+        'source' => {
+          'type' => 'git',
+          'url' => uri.to_s,
+          'rev' => revision,
+          'sha256' => hash,
+          'fetchSubmodules' => submodules
+        }
       }
     end
   end
