@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class Bundix
-  VERSION = "0.0.1" # reseting after fork
+  VERSION = '0.0.1' # reseting after fork
 end
